Added Gen9 pack + animations
I was having trouble with the base gen9 sprites so I pulled the trigger and included the DBK animations as well. If this causes problems in the future, DBK supports toggling off animations or having just a static sprite. Cursory testing shows this as fine, but worst case I should just be able to use a backup of the sprites.